requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
cryptography==3.4.8
redis==3.5.3
requests==2.26.0
python-dotenv==0.19.0 